Templatize the core policy
