+++ CMakeLists.txt
@@ -285,7 +285,6 @@ set(BROKER_SRC
   src/defaults.cc
   src/detail/abstract_backend.cc
   src/detail/clone_actor.cc
-  src/detail/core_policy.cc
   src/detail/data_generator.cc
   src/detail/filesystem.cc
   src/detail/flare.cc
